Move common folder to the components directory Add a key attribute to the link
