--- client-mvp-04/src/common/button/index.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import './index.scss';
-
-export default ({ content, className }) => {
-  return <button className={className}>{content}</button>;
-};
