+++ client-mvp-04/src/components/common/button/index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import './index.scss';
+
+const Button = ({ content, className }) => {
+  return (
+    <button data-testid="button" className={className}>
+      {content}
+    </button>
+  );
+};
+
+export default Button;
