+++ client-mvp-04/src/components/common/button/index.scss
@@ -1,4 +1,4 @@
-@import "../../sass/_main.scss";
+@import "../../../sass/main";
 
 .home-button {
   display: flex;
