+++ client-mvp-04/src/components/header/index.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import './index.scss';
 import logo from '../../assets/images/logo.svg';
-import RedirectLink from '../../common/link';
+import RedirectLink from '../common/link';
 
 const Header = () => {
   return (
     <header className="app-header">
       <RedirectLink
-        key={'header-home'}
+        linkKey={'header-home-link'}
         path={'/'}
         content={<img src={logo} className="app-logo" alt="logo" />}
       />
